Finalize modify usecase

+++ src/main/java/use_case/club/ModifyClubUsecase.java
@@ -4,6 +4,7 @@ import Data.ClubDataAccess;
 import Data.UserDataAccess;
 import Entity.Club;
 
+import javax.swing.*;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -40,14 +41,28 @@ public class ModifyClubUsecase implements ModifyInputBoundary {
         modifyPresenter.setClubList(clubList);
     }
 
+    @Override
+    public void presentDescription(String clubName) {
+        for (Club c : clubDataAccess.getClubs()){
+            if (clubName.equals(c.getName()))
+                modifyPresenter.setDescription(c.getDescription());
+        }
+    }
+
     public void modifyDescription(String clubName, String description) {
-        clubDataAccess.modifyDescription(clubName, description);
-        modifyPresenter.prepareSuccessView("The description is modified!");
+        if (clubName == null)
+            modifyPresenter.prepareFailView("You have not selected any club!");
+        else {
+            clubDataAccess.modifyDescription(clubName, description);
+            modifyPresenter.prepareSuccessView("The description is modified!");
+        }
     }
 
     @Override
     public void deleteClub(String name) {
-        clubDataAccess.deleteClub(name);
-        modifyPresenter.prepareSuccessView("The club is deleted!");
+        if (modifyPresenter.prepareConfirmView("Are you sure you want to delete this club?") == 0) {
+            clubDataAccess.deleteClub(name);
+            modifyPresenter.prepareSuccessView("The club is deleted!");
+        }
     }
 }
